refactor: 패키지 구조 수정, 엔티티 수정
