+++ src/main/java/com/example/wherewego/domain/courses/entity/CourseBookmark.java
@@ -1,7 +1,6 @@
-package com.example.wherewego.domain.courseBookmarks.entity;
+package com.example.wherewego.domain.courses.entity;
 
 import com.example.wherewego.common.entity.BaseEntity;
-import com.example.wherewego.domain.courses.entity.Course;
 import com.example.wherewego.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
