+++ src/main/java/com/example/wherewego/domain/courses/entity/Like.java
@@ -1,7 +1,6 @@
-package com.example.wherewego.domain.placeBookmarks.entity;
+package com.example.wherewego.domain.courses.entity;
 
 import com.example.wherewego.common.entity.BaseEntity;
-import com.example.wherewego.domain.placeReferences.entity.PlaceReference;
 import com.example.wherewego.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -10,23 +9,23 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor
-@Table(name = "place_bookmarks")
-public class PlaceBookmark extends BaseEntity {
+@Table(name = "likes")
+public class Like extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long placeBookmarkId;
+    private Long likeId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "reference_id", nullable = false)
-    private PlaceReference placeReference;
+    @JoinColumn(name = "course_id", nullable = false)
+    private Course course;
 
-    public PlaceBookmark(User user, PlaceReference placeReference) {
+    public Like(User user, Course course) {
         this.user = user;
-        this.placeReference = placeReference;
+        this.course = course;
     }
 }
