+++ src/main/java/com/example/wherewego/domain/places/entity/Place.java
@@ -1,8 +1,10 @@
-package com.example.wherewego.domain.coursePlaces.entity;
+package com.example.wherewego.domain.places.entity;
+
 
 import com.example.wherewego.common.entity.BaseEntity;
 import com.example.wherewego.domain.courses.entity.Course;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,28 +12,25 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor
-@Table(name = "course_places")
-public class CoursePlace extends BaseEntity {
+@AllArgsConstructor
+@Builder
+@Table(name = "places")
+public class Place extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long coursePlaceId;
+    private Long placeId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "course_id", nullable = false)
     private Course course;
 
-    @Column(name = "reference_id", nullable = false)
-    private Long referenceId;
-
     @Column(nullable = false)
     private int visitOrder;
 
-    @Builder
-    public CoursePlace(Course course, Long referenceId, int visitOrder) {
-        this.course = course;
-        this.referenceId = referenceId;
-        this.visitOrder = visitOrder;
-    }
+    @Column(nullable = false, length = 20)
+    private String apiProvider;
 
+    @Column(nullable = false, length = 100)
+    private String apiPlaceId;
 }
